[export][refactor][1/n][reland] Split dynamic shapes
Reland of https://github.com/pytorch/pytorch/pull/114764

[ghstack-poisoned]
